スコア型をi64に変更 荷重標準偏差追加 計算用関数追加

+++ src/app.rs
@@ -184,7 +184,7 @@ impl ScoreTracker {
             return;
         };
 
-        let score = match text.parse::<i32>() {
+        let score = match text.parse::<i64>() {
             Ok(v) => v,
             Err(_) => {
                 self.state.error_message = Some("スコアには整数値を入力してください。".to_string());
