+++ src/models.rs
@@ -20,7 +20,7 @@ pub fn validate_decay_rate_range(rate: f64) -> Result<(), String> {
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct ScoreEntry {
-    pub score: i32,
+    pub score: i64,
     pub timestamp: DateTime<Utc>,
 }
 
@@ -212,7 +212,7 @@ impl AppData {
         &mut self,
         category_name: &str,
         item_name: &str,
-        score: i32,
+        score: i64,
     ) -> Result<(), String> {
         if score < 0 {
             return Err("スコアにマイナスの値は入力できません。".to_string());
